+++ lib/models/text_field_result_list.dart
@@ -6,9 +6,26 @@ class TextFieldResultList {
   TextFieldResultList(this.results);
 
   TextFieldResult find(String id) {
-    return results.firstWhere(
-      (result) => result.id == id,
-      orElse: () => null,
-    );
+    if (contains(id)) {
+      return results.firstWhere(
+        (result) => result.id == id,
+        orElse: () => TextFieldResult(id, ""),
+      );
+    }
+
+    return null;
+  }
+
+  bool contains(String id) {
+    bool exists = false;
+
+    for (final result in results) {
+      if (result.id == id) {
+        exists = true;
+        break;
+      }
+    }
+
+    return exists;
   }
 }