+++ CHANGELOG.md
@@ -30,7 +30,11 @@
 * `TextFormBuilderField` has now the option `obscureText`
 * The `submit` method of `TextFormBuilderState` returns now `TextFieldResultList`
 
-## [1.0.4+4] - 05.10.2020
+## [1.0.4+5] - 05.10.2020
+
+### Added
+* `TextFieldResultList.contains` to check for an id
 
 ### Fixed
 * `TextFieldResultList.find` throws `Bad state: no element`
+* `TextFieldResultList.results = []`